+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.2.1] (2026-05-20)
 ### Changed
 * update crate: flood-tide(0.2.14), flood-tide-gen(0.2.2)
 * to 2024 edition
@@ -18,7 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 * `memx-cdy`
-
 
 ## [0.2.0] (2025-09-15)
 ### Added
@@ -149,7 +151,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] (2021-04-28)
 * first commit
 
-[Unreleased]: https://github.com/aki-akaguma/aki-json-pick/compare/v0.1.10..HEAD
+[Unreleased]: https://github.com/aki-akaguma/aki-json-pick/compare/v0.2.1..HEAD
+[0.2.1]: https://github.com/aki-akaguma/aki-json-pick/compare/v0.2.0..v0.2.1
+[0.2.0]: https://github.com/aki-akaguma/aki-json-pick/compare/v0.1.10..v0.2.0
 [0.1.10]: https://github.com/aki-akaguma/aki-json-pick/compare/v0.1.9..v0.1.10
 [0.1.9]: https://github.com/aki-akaguma/aki-json-pick/compare/v0.1.8..v0.1.9
 [0.1.8]: https://github.com/aki-akaguma/aki-json-pick/compare/v0.1.7..v0.1.8